requirements.txt
fastapi
uvicorn
pandas
polars
yfinance
duckdb
redis
openpyxl
python-dotenv
